// repo/src/types.ts
export interface TelegramBotCommand {
  command: string;
  description: string;
}

export interface TelegramBotMenuButton {
  type: 'commands' | 'web_app' | 'default';
  text?: string;
  url?: string;
}
// ...
export interface TelegramUser {
  id: string;
  firstName: string;
  lastName?: string;
  username?: string;
  phone?: string;
  photoUrl?: string;
  bio?: string;
  isBot?: boolean;
  isVerified?: boolean;
  isPremium?: boolean;
  status?: 'online' | 'offline' | 'recently';
  botInfo?: {
    description?: string;
    about?: string;
    commands?: TelegramBotCommand[];
    menuButton?: TelegramBotMenuButton;
  };
}